Added Tkinter code for modifying rules
Created a UI with a list of boxes that contain all units in the db and can be updated with new ones. These can be dragged and dropped in a set of boxes that are created with the set of rules in the db. This also modifies the database to link the rule and units/remove the link between them.